+++ src/lbfgs.rs
@@ -125,7 +125,7 @@ impl Lbfgs {
         }
     }
 
-    fn search_direction(&self, x: ArrayView1<f64>, g: ArrayView1<f64>) -> Array1<f64> {
+    pub(crate) fn search_direction(&self, x: ArrayView1<f64>, g: ArrayView1<f64>) -> Array1<f64> {
         #[cfg(feature = "highs")]
         if self.highs.is_some() {
             if let Ok(d) = self.highs_step(x, g) {